refactor(tui): move plain data types into app/types.rs (#3314 step 4)
Pure code motion for the App god-object split: the plain data types that
need no App state move from app.rs to the new app/types.rs —
AppMode (+ impl), ReasoningEffort (+ parse/cycle impls and the
work_graph tier conversion), ToolCollapseMode, ComposerDensity,
TranscriptSpacing, VimMode, TuiOptions, InitialInput, QueuedMessage
(+ impl), SubmitDisposition, ToolDetailRecord, TaskPanelEntry,
TaskPanelEntryKind, ApiKeyError, and the AppAction / ShellJobAction /
McpUiAction action enums.

app.rs re-exports every moved item with its original pub visibility, so
all crate::tui::app::X paths (ui.rs, sidebar.rs, setup/, hotbar/, tests,
and the sibling init/composer/status submodules via 'use super::*')
resolve exactly as before. No behavior change, no new public surface.
